add bug wich caused error to showup if not logged in

+++ resources/views/comments/show.blade.php
@@ -95,12 +95,14 @@
                                 @endif
                                 on
                                 {{$c->created_at}}
+                                @if(isset(Auth::user()->name))
                                     @if(Auth::user()->name == $c->name)
-                                    <input type="hidden" name="_token" value="{{ csrf_token() }}">
-                                    <a href="/comments/edit/{{$c->id}}" class="btn btn-primary btn-xs edit-btn">edit</a>
-                                    <a href="/comments/delete/{{$c->id}}" class="btn btn-danger btn-xs edit-btn">
-                                      <i class="fa fa-btn fa-trash" title="delete"></i> delete
-                                    </a>
+                                      <input type="hidden" name="_token" value="{{ csrf_token() }}">
+                                      <a href="/comments/edit/{{$c->id}}" class="btn btn-primary btn-xs edit-btn">edit</a>
+                                      <a href="/comments/delete/{{$c->id}}" class="btn btn-danger btn-xs edit-btn">
+                                        <i class="fa fa-btn fa-trash" title="delete"></i> delete
+                                      </a>
+                                    @endif
                                 @endif
                               </div>
                             </li>
